Fixes Seeder bug
