--- src/database/seeds/BantenprovProgramKeahlianSeeder.csv
@@ -1,2 +1,0 @@
-IPA,ilmu pengetahuan alam,1
-IPS,ilmu pengetahuan sosial,1
