+++ src/database/seeds/BantenprovProgramKeahlianSeederProgramKeahlian.csv
@@ -1,0 +1,5 @@
+1,Teknik Komputer dan Jaringan,Teknik Komputer dan Jaringan,1
+2,Multimedia,Multimedia,1
+3,Rekayasa Perangkat Lunak,Rekayasa Perangkat Lunak,1
+4,Perhotelan,Perhotelan,1
+5,Tata Boga,Tata Boga,1
